% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_podaac}
\alias{archive_podaac}
\title{Archive a PODAAC stars object to a database}
\usage{
archive_podaac(x, orig_filename, path = ".")
}
\arguments{
\item{x}{stars object}

\item{orig_filename}{chr the downloaded PODAAC filename (ends with .nc)}

\item{path}{char, the data path to write the data}
}
\value{
tabular database as a tibble
}
\description{
Archive a PODAAC stars object to a database
}
