% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{compose_filename}
\alias{compose_filename}
\title{Compose a filename from a archive database}
\usage{
compose_filename(x, path = ".", ext = ".tif")
}
\arguments{
\item{x}{database (tibble), with date, var, depth}

\item{path}{character, the root path for the filename}

\item{ext}{character, the filename extension to apply (please include the dot)}
}
\value{
character vector of filenames in form
        \code{<path>/YYYY/mm/datetime-rdac-level-type-product-region-gds_version-file_version_var.ext}
}
\description{
Compose a filename from a archive database
}
