% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_filename}
\alias{decompose_filename}
\title{Decompose an archived filename (.tif) into a database.}
\usage{
decompose_filename(
  x = c("20200201090000-JPL-L4_GHRSST-SSTfnd-MUR-GLOB-v02.0-fv04.1-analyzed_sst.tif",
    "20070503120000-UKMO-L4_GHRSST-SSTfnd-OSTIA-GLOB-v02.1-fv01.0-mask.tif"),
  ext = ".tif"
)
}
\arguments{
\item{x}{character, vector of one or more filenames}

\item{ext}{character, the extension to remove (including dot)}
}
\value{
table (tibble) database
}
\description{
Not to be confused with \link{decompose_podaac_filename}
}
