% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_podaac_filename}
\alias{decompose_podaac_filename}
\title{Decompose a PODAAC original filename into a database of variables.}
\usage{
decompose_podaac_filename(
  x = c("20200201090000-JPL-L4_GHRSST-SSTfnd-MUR-GLOB-v02.0-fv04.1.nc",
    "20070503120000-UKMO-L4_GHRSST-SSTfnd-OSTIA-GLOB-v02.1-fv01.0.nc"),
  vars = c("analysed_sst", "analysis_error", "mask", "sea_ice_fraction", "dt_1km_data",
    "sst_anomaly")
)
}
\arguments{
\item{x}{chr one or more filenames}

\item{vars}{chr or stars object,  The names of the variables in the file. If a stars object we'll
take the variable names from that.}
}
\value{
a database tibble
}
\description{
Here we follow [this](https://isi-sbx.ifremer.fr/jeff/gds/naming.html) nice
documentation of the naming convetion.
}
