% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{.extract_point}
\alias{.extract_point}
\title{Extract points for a given variable}
\usage{
.extract_point(tbl, key, X = NULL)
}
\arguments{
\item{tbl}{table of navigation info, see \code{\link{mur_nc_nav_point}}}

\item{key}{table of variable name}

\item{X}{\code{ncdf4} object}
}
\value{
table of variable values
}
\description{
Extract points for a given variable
}
