% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{extract generic}
\usage{
extract(x, y, ...)
}
\arguments{
\item{x}{\code{sf} object}

\item{y}{\code{ncdf4} object}

\item{...}{Arguments passed to or from other methods}
}
\value{
data frame of covariates (point, or raster)
}
\description{
extract generic
}
