% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podaac.R
\name{format_date}
\alias{format_date}
\title{Format a date-time string}
\usage{
format_date(x, fmt = "\%Y-\%m-\%dT00:00:00Z")
}
\arguments{
\item{x}{a Date or POSIXt object}

\item{fmt}{chr, the desired output format}
}
\value{
character formatted date stamp
}
\description{
Format a date-time string
}
