% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{get_namibia}
\alias{get_namibia}
\title{Retrieve a coordinates to define offshore Namibia}
\usage{
get_namibia(form = c("table", "sf", "bbox")[2])
}
\arguments{
\item{form}{chr, one of 'table', 'sf' or 'bbox' to define output form}
}
\value{
tibble, sf or bbox class object
}
\description{
Retrieve a coordinates to define offshore Namibia
}
