% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghrsst-package.R
\docType{package}
\name{ghrsst}
\alias{ghrsst}
\title{ghrsst: Simplified access to GHRSST (JPL's MUR daily SST) data from R.}
\description{
Simplified access to GHRSST (JPL's MUR daily SST) data from R.
}
