% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mur.R
\name{mur_url}
\alias{mur_url}
\title{Craft a MUR URL for a given date}
\usage{
mur_url(
  date = Sys.Date() - 2,
  where = "opendap",
  root = file.path("https://opendap.jpl.nasa.gov",
    "opendap/OceanTemperature/ghrsst/data/GDS2/L4", "GLOB/JPL/MUR/v4.1"),
  product = "JPL-L4_GHRSST-SSTfnd-MUR-GLOB-v02.0-fv04.1.nc"
)
}
\arguments{
\item{date}{character, POSIXt or Date the date to retrieve}

\item{where}{character ignored (for now)}

\item{root}{character, the root URL}

\item{product}{character, provides version and extend info, leave as default}
}
\value{
one or more URLs
}
\description{
Craft a MUR URL for a given date
}
