% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podaac.R
\name{podaac_date}
\alias{podaac_date}
\title{Extract a date form a PODAAC filename}
\usage{
podaac_date(filename)
}
\arguments{
\item{filename}{chr, one or more PODAAC filenames}
}
\value{
a date for each filename
}
\description{
Extract a date form a PODAAC filename
}
