% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podaac.R
\name{podaac_list}
\alias{podaac_list}
\title{List files stored in the PODAAC temporary directory}
\usage{
podaac_list(path = ghrsst_path("tmp"), pattern = "^.*\\\\.nc$")
}
\arguments{
\item{path}{char, the data path}

\item{pattern}{char one or more regex patterns for the file search}
}
\value{
char file listing
}
\description{
List files stored in the PODAAC temporary directory
}
