% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podaac.R
\name{purge_podaac}
\alias{purge_podaac}
\title{Delete files stored in the PODAAC temporary directory}
\usage{
purge_podaac(
  path = ghrsst_path("tmp"),
  logfile = file.path(path, "purge-log"),
  verbose = FALSE,
  pattern = "^.*\\\\.nc$"
)
}
\arguments{
\item{path}{char, the data path}

\item{logfile}{char the name of the log file}

\item{verbose}{logical, output messages if TRUE}

\item{pattern}{char one or more regex patterns for the file search}
}
\value{
named logical vector, TRUE indicates the file was successfully deleted
}
\description{
Delete files stored in the PODAAC temporary directory
}
