% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{read_ghrsst}
\alias{read_ghrsst}
\title{Read GHRSST data as stars}
\usage{
read_ghrsst(db, path)
}
\arguments{
\item{db}{tibble database}

\item{path}{chr the data path}
}
\value{
a stars object
}
\description{
Read GHRSST data as stars
}
