% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{read_podaac}
\alias{read_podaac}
\title{Read one or more PODAAC files as stars objects.  
 
You can extract the entore globe, but these are large files and you may be
happier to select a subset using the `bb` argument. If multiple filenames are
passed then}
\usage{
read_podaac(filename, bb = get_namibia(), var = "analysed_sst")
}
\arguments{
\item{filename}{chr one or more filenames}

\item{bb}{bbox or object from which a bbox can be extracted}

\item{var}{char, one or more variables to return. By default 'analysed_sst', but
there is also 'analysis_error', 'mask', 'sea_ice_fraction', 'dt_1km_data', and
'sst_anomaly'}
}
\value{
stars object with one or variables and one or more bands
}
\description{
Read one or more PODAAC files as stars objects.  
 
You can extract the entore globe, but these are large files and you may be
happier to select a subset using the `bb` argument. If multiple filenames are
passed then
}
