% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{tidync_as_stars}
\alias{tidync_as_stars}
\title{Cast a tidync object to stars}
\usage{
tidync_as_stars(x, var = "analysed_sst")
}
\arguments{
\item{x}{tidync object}

\item{var}{char, one or more variables to return. By default 'analysed_sst', but
there is also 'analysis_error', 'mask', 'sea_ice_fraction', 'dt_1km_data', and
'sst_anomaly'}
}
\value{
stars object
}
\description{
Taken from https://github.com/ropensci/tidync/issues/68#issuecomment-484773118
}
