% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{write_database}
\alias{write_database}
\title{Write the file-list database}
\usage{
write_database(x, path, filename = "database.csv.gz")
}
\arguments{
\item{x}{the tibble or data.frame database}

\item{path}{character the directory to where the database should reside}

\item{filename}{character, optional filename}
}
\value{
the input tibble
}
\description{
We save only date (YYYY-mm-dd), param, trt (treatment) and src (source). If you
have added other variables to the database they will be dropped in the saved
file.
}
